// src/utils/interfaces/Pokemon/Pokemon.ts
export interface IAllPokemonsResults {
  name: string;
  url: string;
}
export interface IAllPokemons {
  count: number;
  next: string | null;
  previous: string | null;
  results: IAllPokemonsResults[];
}

export interface IPokemon {
  id: number;
  name: string;
  base_experience: number;
  height: number;
  is_default: boolean;
  order: number;
  weight: number;
  abilities: IPokemonAbility[];
  forms: Array<INamedApiResource<IPokemonForm>>;
  game_indices: IVersionGameIndex[];
  held_items: IPokemonHeldItem[];
  location_area_encounters: string;
  moves: IPokemonMove[];
  sprites: IPokemonSprites;
  species: INamedApiResource<IPokemonSpecies>;
  stats: IPokemonStat[];
  types: IPokemonType[];
}

export interface IPokemonAbility {
  is_hidden: true;
  slot: number;
  ability: INamedApiResource<IAbility>;
}

export interface IPokemonType {
  slot: number;
  type: INamedApiResource<IType>;
}

export interface IPokemonHeldItem {
  item: INamedApiResource<IItem>;
  version_details: IPokemonHeldItemVersion[];
}

export interface IPokemonHeldItemVersion {
  version: INamedApiResource<IVersion>;
  rarity: number;
}

export interface IPokemonMove {
  move: INamedApiResource<IMove>;
  version_group_details: IPokemonMoveVersion[];
}

export interface IPokemonMoveVersion {
  move_learn_method: INamedApiResource<IMoveLearnMethod>;
  version_group: INamedApiResource<IVersionGroup>;
  level_learned_at: number;
}

export interface IPokemonStat {
  stat: INamedApiResource<IStat>;
  effort: number;
  base_stat: number;
}

export interface IPokemonSprites {
  back_default: string;
  back_female: string;
  back_shiny: string;
  back_shiny_female: string;
  front_default: string;
  front_female: string;
  front_shiny: string;
  front_shiny_female: string;
  other: IPokemonSpriteOther;
  versions: IPokemonSpriteVersion;
}

interface ISpriteVariant {
  back_default: string | null;
  back_female: string | null;
  back_gray: string | null;
  back_shiny: string | null;
  back_shiny_female: string | null;
  front_default: string | null;
  front_female: string | null;
  front_gray: string | null;
  front_shiny: string | null;
  front_shiny_female: string | null;
}

export interface IPokemonSpriteOther {
  dream_world: Pick<ISpriteVariant, "front_default" | "front_female">;
  "official-artwork": {
    front_default: string;
  };
}
// ...
interface IPokemonSpriteVersion {
  "generation-i": IGeneration1Sprite;
  "generation-ii": IGeneration2Sprite;
  "generation-iii": IGeneration3Sprite;
  "generation-iv": IGeneration4Sprite;
  "generation-v": IGeneration5Sprite;
  "generation-vi": IGeneration6Sprite;
  "generation-vii": IGeneration7Sprite;
  "generation-viii": IGeneration8Sprite;
}